+++ packages/hex/types/index.d.ts
@@ -4,7 +4,7 @@ declare enum HexValue {}
 
 export type hexValue = HexValue & string
 
-export type hexInput = hexValue | string | Object | Buffer
+export type hexInput = hexValue | string | Object | Buffer | Array<number>
 
 export function isHexValue(input: any): boolean
 
@@ -12,6 +12,7 @@ export function fromHexValue(input: hexValue): Buffer
 
 type HexInput =
   | { type: 'buffer'; value: Buffer }
+  | { type: 'bytesArray'; value: Array<number> }
   | { type: 'hex'; value: hexValue }
   | { type: 'object'; value: Object }
   | { type: 'string'; value: string }
@@ -21,6 +22,7 @@ export class Hex {
   value: hexValue
   equals(other: hexInput | Hex): boolean
   toBuffer(): Buffer
+  toBytesArray(): Array<number>
   toObject(): Object
   toString(): string
 }