+++ packages/secp256k1/types/index.d.ts
@@ -3,12 +3,29 @@
 import BN = require('bn.js')
 import elliptic = require('elliptic')
 
-export function createKeyPair(
-  privKey?: string,
-  enc?: string,
-): elliptic.ec.KeyPair
+export interface SignatureBuffer {
+  r: Buffer
+  s: Buffer
+}
+
+export interface SignatureHex {
+  r: string
+  s: string
+}
+
+export type Signature = SignatureBuffer | SignatureHex
+
+export function createKeyPair(privKey?: string): elliptic.ec.KeyPair
+
+export function createPublic(pubKey: string): elliptic.ec.KeyPair
 
 export function sign(
-  digest: Array<number>,
+  bytes: Array<number>,
   privKey: Buffer | BN | string,
 ): Array<number>
+
+export function verify(
+  bytes: Array<number>,
+  signature: Array<number> | Signature,
+  pubKey: string | elliptic.ec.KeyPair,
+): boolean